Optimized for Sumo Service Region
Modified `New-SumoCredential` to detect the service region associated
with the user’s Sumo Logic instance and write that value out to a file
to allow reuse, along with the user’s credentials.

This solves an issue where HTTP 401 redirects result in a HTTP 401
Unauthorized message being returned because the `Invoke-RestMethod`
cmdlet does not send Basic Authentication headers upon redirection. By
ensuring that future requests go directly to the correct URI and are
not redirected, this problem is avoided.

Also modified the other functions to use the global variable containing
the correct URI rather than a hardcoded value.