Fix HTTP 403 Issue
Added logic to detect the correct Sumo Logic service region to use
during the initial credential validation in `New-SumoCredential` cmdlet.
The `$global:SumoBaseAPIHost` variable is updated with this information,
and is saved in a text file for future sessions.

This avoids HTTP 403 authentication issues on redirect, since the Sumo
API will not redirect requests that are sent to the correct service
region to begin with.